+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 # Changelog
-## 2.3.0-SNAPSHOT
+## 2.3.0
+- Added support for blocks with upper-diagonal blocks like acacia or dark oak
+- Added support for crimson and warped fungus trees
+- Removed tree-types config option. Let me know if you need this again
 - Added Spanish and Turkish translation